Functional Worker Integration tests
